requirements.txt
# numpy==1.20.3 Example requirement
beautifulsoup4==4.10.0
requests==2.26.0
git+https://github.com/huggingface/transformers.git
torch==1.9.0
google-cloud-storage==1.42.1